course-mcq-generator/backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.34.0
pydantic==2.11.2
langchain==0.3.20
langchain-google-genai==2.1.10
langchain-community==0.3.19
langchain-core==0.3.41
langchain-text-splitters==0.3.6
faiss-cpu==1.8.0.post1
PyMuPDF==1.26.4
pdfplumber==0.11.2
python-multipart==0.0.12
python-dotenv==1.0.1
aiofiles==24.1.0